Refactor: LoadBlockIndex() code movement.
